# fare-payment-service/postman/fare-payment-openapi.json
{"openapi":"3.1.0","info":{"title":"RideLink - Fare & Payment Service","description":"Fare estimation, final fare calculation, simulated payments and receipts. All payments are simulated; no real payment gateway is used.","version":"1.0.0"},"servers":[{"url":"http://localhost:8084","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Fares","description":"Fare estimation and final fare calculation"},{"name":"Payments","description":"Simulated payments, payment status and receipts"}],"paths":{"/api/payments":{"post":{"tags":["Payments"],"summary":"Record a simulated payment for a ride","description":"The amount always comes from the stored final fare. The response status is COMPLETED (receipt issued) or FAILED (when simulateFailure is true). A FAILED payment can be retried. Roles: PASSENGER, ADMIN.","operationId":"pay","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentRequest"}}},"required":true},"responses":{"201":{"description":"Payment recorded (COMPLETED or FAILED)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"400":{"description":"Invalid input (blank rideId or unknown method)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"401":{"description":"Missing or invalid token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"403":{"description":"Role not allowed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"404":{"description":"No final fare exists for this ride","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"409":{"description":"Ride has already been paid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}},"/api/fares/final":{"post":{"tags":["Fares"],"summary":"Calculate and store the final fare of a ride","description":"Called when a ride is completed. Provide actualDistanceKm, or both pickup and destination to calculate the distance. One final fare per ride. Roles: DRIVER, ADMIN.","operationId":"calculateFinalFare","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FinalFareRequest"}}},"required":true},"responses":{"201":{"description":"Final fare stored","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}},"400":{"description":"Invalid input (blank ids, negative distance, unknown location, or no distance/locations given)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}},"401":{"description":"Missing or invalid token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}},"403":{"description":"Role not allowed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}},"409":{"description":"A final fare already exists for this ride","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}}}}},"/api/fares/estimate":{"post":{"tags":["Fares"],"summary":"Estimate a fare","description":"Calculates the fare for a pickup and destination (known place names or \"lat,lon\" coordinates) using fare = baseFare + distanceKm x ratePerKm, with a minimum fare. Nothing is stored. No token needed: this is the internal call made by Ride Management.","operationId":"estimate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FareEstimateRequest"}}},"required":true},"responses":{"200":{"description":"Estimate calculated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareEstimateResponse"}}}},"400":{"description":"Missing pickup or destination, or unknown location","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareEstimateResponse"}}}}}}},"/api/payments/{paymentId}":{"get":{"tags":["Payments"],"summary":"Get a payment and its status","description":"Any authenticated user.","operationId":"getPayment","parameters":[{"name":"paymentId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Payment found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"401":{"description":"Missing or invalid token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"404":{"description":"Payment not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}},"/api/payments/{paymentId}/receipt":{"get":{"tags":["Payments"],"summary":"Get the receipt of a completed payment","description":"Any authenticated user.","operationId":"getReceipt","parameters":[{"name":"paymentId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Receipt generated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReceiptResponse"}}}},"401":{"description":"Missing or invalid token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReceiptResponse"}}}},"404":{"description":"Payment not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReceiptResponse"}}}},"409":{"description":"Payment is not COMPLETED","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReceiptResponse"}}}}}}},"/api/payments/ride/{rideId}":{"get":{"tags":["Payments"],"summary":"List payment attempts for a ride, newest first","description":"Shows failed attempts, retries and the completed payment. Any authenticated user.","operationId":"getPaymentsByRide","parameters":[{"name":"rideId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Payment attempts (empty list if none)","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentResponse"}}}}},"401":{"description":"Missing or invalid token","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}}},"/api/fares/ride/{rideId}":{"get":{"tags":["Fares"],"summary":"Get the stored final fare of a ride","description":"Any authenticated user.","operationId":"getFareByRideId","parameters":[{"name":"rideId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Fare found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}},"401":{"description":"Missing or invalid token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}},"404":{"description":"No fare stored for this ride","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FareResponse"}}}}}}}},"components":{"schemas":{"PaymentRequest":{"type":"object","properties":{"rideId":{"type":"string","minLength":1},"method":{"type":"string","enum":["CASH","CARD"]},"simulateFailure":{"type":"boolean"}},"required":["method","rideId"]},"PaymentResponse":{"type":"object","properties":{"id":{"type":"string"},"rideId":{"type":"string"},"passengerId":{"type":"string"},"amount":{"type":"number"},"currency":{"type":"string"},"method":{"type":"string","enum":["CASH","CARD"]},"status":{"type":"string","enum":["PENDING","COMPLETED","FAILED"]},"receiptNumber":{"type":"string"},"failureReason":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"paidAt":{"type":"string","format":"date-time"}}},"FinalFareRequest":{"type":"object","properties":{"rideId":{"type":"string","minLength":1},"passengerId":{"type":"string","minLength":1},"actualDistanceKm":{"type":"number","minimum":0.0},"pickup":{"type":"string"},"destination":{"type":"string"}},"required":["passengerId","rideId"]},"FareResponse":{"type":"object","properties":{"id":{"type":"string"},"rideId":{"type":"string"},"passengerId":{"type":"string"},"distanceKm":{"type":"number"},"baseFare":{"type":"number"},"ratePerKm":{"type":"number"},"totalFare":{"type":"number"},"currency":{"type":"string"},"calculatedAt":{"type":"string","format":"date-time"}}},"FareEstimateRequest":{"type":"object","properties":{"pickup":{"type":"string","minLength":1},"destination":{"type":"string","minLength":1}},"required":["destination","pickup"]},"FareEstimateResponse":{"type":"object","properties":{"pickup":{"type":"string"},"destination":{"type":"string"},"distanceKm":{"type":"number"},"baseFare":{"type":"number"},"ratePerKm":{"type":"number"},"minimumFare":{"type":"number"},"estimatedFare":{"type":"number"},"currency":{"type":"string"}}},"ReceiptResponse":{"type":"object","properties":{"receiptNumber":{"type":"string"},"paymentId":{"type":"string"},"rideId":{"type":"string"},"passengerId":{"type":"string"},"distanceKm":{"type":"number"},"baseFare":{"type":"number"},"ratePerKm":{"type":"number"},"totalPaid":{"type":"number"},"currency":{"type":"string"},"method":{"type":"string","enum":["CASH","CARD"]},"paidAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}